Generalize VAST plugin for all ZeroMQ apps
Rename the plugin to threatbus-zmq-app
Rename package to threatbus_zmq_app
Adjust all configs, Makefile, Dockerfile
Extract message mapping, use JSON marshalling of native Threat Bus types
