+++ Makefile
@@ -43,7 +43,7 @@ build:
 	python setup.py build
 	python plugins/apps/threatbus_zeek/setup.py build
 	python plugins/apps/threatbus_misp/setup.py build
-	python plugins/apps/threatbus_vast/setup.py build
+	python plugins/apps/threatbus_zmq_app/setup.py build
 	python plugins/apps/threatbus_cif3/setup.py build
 	python plugins/backbones/threatbus_inmem/setup.py build
 	python plugins/backbones/threatbus_rabbitmq/setup.py build
@@ -56,7 +56,7 @@ dist:
 	make clean
 	python plugins/apps/threatbus_misp/setup.py sdist bdist_wheel
 	make clean
-	python plugins/apps/threatbus_vast/setup.py sdist bdist_wheel
+	python plugins/apps/threatbus_zmq_app/setup.py sdist bdist_wheel
 	make clean
 	python plugins/apps/threatbus_cif3/setup.py sdist bdist_wheel
 	make clean
@@ -69,7 +69,7 @@ install:
 	python setup.py install
 	python plugins/apps/threatbus_zeek/setup.py install
 	python plugins/apps/threatbus_misp/setup.py install
-	python plugins/apps/threatbus_vast/setup.py install
+	python plugins/apps/threatbus_zmq_app/setup.py install
 	python plugins/apps/threatbus_cif3/setup.py install
 	python plugins/backbones/threatbus_inmem/setup.py install
 	python plugins/backbones/threatbus_rabbitmq/setup.py install
@@ -79,7 +79,7 @@ dev-mode:
 	python setup.py develop
 	python plugins/apps/threatbus_zeek/setup.py develop
 	python plugins/apps/threatbus_misp/setup.py develop
-	python plugins/apps/threatbus_vast/setup.py develop
+	python plugins/apps/threatbus_zmq_app/setup.py develop
 	python plugins/apps/threatbus_cif3/setup.py develop
 	python plugins/backbones/threatbus_inmem/setup.py develop
 	python plugins/backbones/threatbus_rabbitmq/setup.py develop
