+++ apps/vast/requirements.txt
@@ -1,3 +1,4 @@
 coloredlogs>=14.0
 pyzmq>=19
 pyvast>=2020.8.28
+threatbus>=2020.9.30
