+++ config.yaml.example
@@ -45,12 +45,11 @@ plugins:
       #    bootstrap.servers: "localhost:9092"
       #    group.id: "threatbus"
       #    auto.offset.reset: "earliest"
-    vast:
-      zmq:
-        host: "127.0.0.1"
-        manage: 13370
-        pub: 13371
-        sub: 13372
+    zmq_app:
+      host: "127.0.0.1"
+      manage: 13370
+      pub: 13371
+      sub: 13372
     cif3:
       api:
         host: http://localhost:5000
