+++ plugins/apps/threatbus_zmq_app/setup.py
@@ -23,8 +23,8 @@ setup(
         "Topic :: Software Development :: Object Brokering",
         "Topic :: System :: Distributed Computing",
     ],
-    description="A plugin to enable threatbus communication with VAST.",
-    entry_points={"threatbus.app": ["vast = threatbus_vast.plugin"]},
+    description="A plugin to connect apps via ZeroMQ.",
+    entry_points={"threatbus.app": ["zmq-app = threatbus_zmq_app.plugin"]},
     install_requires=[
         "threatbus>=2020.09.30",
         "pyzmq>=19",
@@ -32,16 +32,17 @@ setup(
     ],
     keywords=[
         "threatbus",
-        "vast",
         "threat intelligence",
+        "plugin",
         "zeromq",
+        "zmq",
     ],
     license="BSD 3-clause",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    name="threatbus-vast",
+    name="threatbus-zmq-app",
     package_dir={"": "plugins/apps"},
-    packages=["threatbus_vast"],
+    packages=["threatbus_zmq_app"],
     python_requires=">=3.7",
     url="https://github.com/tenzir/threatbus",
     version="2020.09.30",
